Move method to pipeline
